+++ mobile/android/android-components/components/browser/session/src/test/java/mozilla/components/browser/session/engine/EngineObserverTest.kt
@@ -68,7 +68,12 @@ class EngineObserverTest {
                 notifyObservers { onLoadingStateChange(true) }
                 notifyObservers { onNavigationStateChange(true, true) }
             }
-            override fun loadUrl(url: String, parent: EngineSession?, flags: LoadUrlFlags) {
+            override fun loadUrl(
+                url: String,
+                parent: EngineSession?,
+                flags: LoadUrlFlags,
+                additionalHeaders: Map<String, String>?
+            ) {
                 notifyObservers { onLocationChange(url) }
                 notifyObservers { onProgress(100) }
                 notifyObservers { onLoadingStateChange(true) }
@@ -110,7 +115,12 @@ class EngineObserverTest {
             override fun saveState(): EngineSessionState = mock()
             override fun loadData(data: String, mimeType: String, encoding: String) {}
             override fun recoverFromCrash(): Boolean { return false }
-            override fun loadUrl(url: String, parent: EngineSession?, flags: LoadUrlFlags) {
+            override fun loadUrl(
+                url: String,
+                parent: EngineSession?,
+                flags: LoadUrlFlags,
+                additionalHeaders: Map<String, String>?
+            ) {
                 if (url.startsWith("https://")) {
                     notifyObservers { onSecurityChange(true, "host", "issuer") }
                 } else {
@@ -148,7 +158,12 @@ class EngineObserverTest {
 
             override fun toggleDesktopMode(enable: Boolean, reload: Boolean) {}
             override fun saveState(): EngineSessionState = mock()
-            override fun loadUrl(url: String, parent: EngineSession?, flags: LoadUrlFlags) {}
+            override fun loadUrl(
+                url: String,
+                parent: EngineSession?,
+                flags: LoadUrlFlags,
+                additionalHeaders: Map<String, String>?
+            ) {}
             override fun loadData(data: String, mimeType: String, encoding: String) {}
             override fun findAll(text: String) {}
             override fun findNext(forward: Boolean) {}